+++ data/scripting/pickup.lua
@@ -1,6 +1,0 @@
-local raw = trxc.pickup
-local Mode = trxc.pickup.Mode
-
-local pickup = { Mode = Mode }
-
-trx.pickup = pickup