+++ data/trx/ship/games/tr4/scripts/csplit1.lua
@@ -1,8 +1,8 @@
 trx.events.before_item_setup(function(level)
-  trx.items[59].properties.pickup_mode = trx.pickup.Mode.PLINTH_LOW
-  trx.items[93].properties.pickup_mode = trx.pickup.Mode.PLINTH_LOW
-  trx.items[105].properties.pickup_mode = trx.pickup.Mode.PLINTH_LOW
-  trx.items[121].properties.pickup_mode = trx.pickup.Mode.PLINTH_LOW
+  trx.items[59].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
+  trx.items[93].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
+  trx.items[105].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
+  trx.items[121].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
 
   trx.objects.animating_16.properties.collidable = false
 end)