+++ data/trx/ship/games/tr4/scripts/libend.lua
@@ -1,3 +1,3 @@
 trx.events.before_item_setup(function(level)
-  trx.items[14].properties.pickup_mode = trx.pickup.Mode.PLINTH_LOW
+  trx.items[14].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
 end)