+++ data/trx/ship/games/tr4/scripts/palaces.lua
@@ -1,4 +1,4 @@
 trx.events.before_item_setup(function(level)
-  trx.items[64].properties.pickup_mode = trx.pickup.Mode.PLINTH_LOW
-  trx.items[105].properties.pickup_mode = trx.pickup.Mode.PLINTH_LOW
+  trx.items[64].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
+  trx.items[105].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
 end)