+++ src/trx/game/enum.c
@@ -7,6 +7,7 @@
 #include <trx/game/items/enum.h>
 #include <trx/game/lara/skin/types.h>
 #include <trx/game/lara/types.h>
+#include <trx/game/objects/general/pickup.h>
 #include <trx/game/objects/ids.h>
 #include <trx/game/rooms/enum.h>
 #include <trx/game/screenshot.h>
@@ -63,6 +64,10 @@ static __attribute__((constructor)) void M_Init(void)
     ENUM_MAP(ITEM_STATUS, IS_ACTIVE, "active");
     ENUM_MAP(ITEM_STATUS, IS_DEACTIVATED, "deactivated");
     ENUM_MAP(ITEM_STATUS, IS_INVISIBLE, "invisible");
+
+    ENUM_MAP(PICKUP_MODE, PICKUP_MODE_NORMAL, "normal");
+    ENUM_MAP(PICKUP_MODE, PICKUP_MODE_PLINTH_LOW, "plinth_low");
+    ENUM_MAP(PICKUP_MODE, PICKUP_MODE_PLINTH_HIGH, "plinth_high");
 
     ENUM_MAP(ROOM_FLIP_STATUS, RFS_NONE, "none");
     ENUM_MAP(ROOM_FLIP_STATUS, RFS_UNFLIPPED, "unflipped");